requirements.txt
# Core dependencies for arXiv Aggregator
feedparser>=6.0.10
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0

# Optional dependencies (uncomment if needed)
# beautifulsoup4>=4.12.0
# lxml>=4.9.0